feat(wave1): wire svelte2tsx into compatibility report dashboard
Wave 1 acceptance criteria from docs/ecosystem-implementation-plan.md:

- [x] cargo test --release --test svelte2tsx_fixtures reports 100% (245/245)
- [x] napi_svelte2tsx unchanged
- [x] svelte2tsx wired into the compatibility report dashboard

Move the runner from a closed-over module in tests/svelte2tsx_fixtures.rs
into tests/common/svelte2tsx.rs so the standalone test and the
compatibility report category share one code path. The dashboard now
shows svelte2tsx as a category alongside print, preprocess, parser,
runtime, etc.

Side effect: the standalone runner no longer needs the
#[cfg(not(feature = "native"))] gate — it builds under any feature
combination now.

Compatibility report total: 3341/3341 in-scope passing (was 3096/3096
+ 245 new svelte2tsx fixtures).
