+++ tests/svelte2tsx_fixtures.rs
@@ -1,1034 +1,94 @@
-//! Integration tests for svelte2tsx against language-tools test fixtures.
-//!
-//! These tests require:
-//!   1. The language-tools submodule to be checked out
-//!   2. The `native` feature to be disabled (svelte2tsx is not compiled with `native`)
+//! Standalone svelte2tsx fixture runner. The actual logic lives in
+//! `tests/common/svelte2tsx.rs` so the same code drives the compatibility
+//! report dashboard's svelte2tsx category.
 //!
 //! Run with:
-//!   cargo test --no-default-features --test svelte2tsx_fixtures -- --nocapture
+//!   cargo test --test svelte2tsx_fixtures -- --nocapture
 //!
-//! The test prints a summary of pass/fail/skip counts and the first differing
-//! lines for each failing sample.
-
-#[cfg(not(feature = "native"))]
-mod svelte2tsx_tests {
-    use std::fs;
-    use std::path::{Path, PathBuf};
-
-    use svelte_compiler_rust::svelte2tsx::{
-        RewriteExternalImportsOptions, Svelte2TsxMode, Svelte2TsxNamespace, Svelte2TsxOptions,
-        SvelteVersion, svelte2tsx,
-    };
-
-    // =========================================================================
-    // Helpers
-    // =========================================================================
-
-    /// Normalize line endings and trim trailing whitespace (matches JS `normalize` helper).
-    fn normalize(content: &str) -> String {
-        content.replace("\r\n", "\n").trim_end().to_string()
-    }
-
-    /// Find the first `.svelte` file in a sample directory.
-    /// Most samples use `input.svelte`, but some have custom names
-    /// (e.g. `+page.svelte`, `0.svelte`).
-    fn find_svelte_file(sample_dir: &Path) -> Option<PathBuf> {
-        let mut entries: Vec<_> = fs::read_dir(sample_dir)
-            .ok()?
-            .filter_map(|e| e.ok())
-            .filter(|e| e.path().extension().is_some_and(|ext| ext == "svelte"))
-            .collect();
-        // Sort for determinism (prefer `input.svelte` if multiple exist)
-        entries.sort_by_key(|e| e.file_name());
-        entries.into_iter().next().map(|e| e.path())
-    }
-
-    /// Build svelte2tsx options from the sample name.
-    ///
-    /// This mirrors the JS `get_svelte2tsx_config` function:
-    /// - `ts-*` samples set `is_ts_file: true`
-    /// - `*-dts` samples set `mode: Dts`
-    /// - `accessors-config*` samples set `accessors: true`
-    /// - `*-foreign-ns` samples should set namespace to `Foreign`
-    ///   (not yet available in Rust enum, defaults to Html)
-    /// - `config.json` in the sample dir can override `filename`
-    fn build_options(
-        sample_name: &str,
-        sample_dir: &Path,
-        svelte_filename: &str,
-    ) -> Svelte2TsxOptions {
-        let is_ts_file = sample_name.starts_with("ts-");
-
-        let mode = if sample_name.ends_with("-dts") {
-            Svelte2TsxMode::Dts
-        } else {
-            Svelte2TsxMode::Ts
-        };
-
-        let accessors = sample_name.starts_with("accessors-config");
-
-        // NOTE: The JS test sets namespace to 'foreign' for *-foreign-ns samples.
-        // Our Rust enum does not have a Foreign variant yet, so we default to Html.
-        let namespace = Svelte2TsxNamespace::Html;
-
-        let version = SvelteVersion::V5;
-
-        // `jsdoc-*` samples enable JSDoc emit format (matching JS test runner)
-        let emit_jsdoc = sample_name.starts_with("jsdoc-") || sample_name.starts_with("js-jsdoc-");
-
-        // Read config.json overrides if present
-        let mut filename = svelte_filename.to_string();
-        let mut rewrite_external_imports: Option<RewriteExternalImportsOptions> = None;
-        let config_path = sample_dir.join("config.json");
-        if config_path.exists() {
-            if let Ok(config_str) = fs::read_to_string(&config_path) {
-                if let Ok(config) = serde_json::from_str::<serde_json::Value>(&config_str) {
-                    if let Some(f) = config.get("filename").and_then(|v| v.as_str()) {
-                        filename = f.to_string();
-                    }
-                    if let Some(rew) = config.get("rewriteExternalImports") {
-                        let workspace = rew
-                            .get("workspacePath")
-                            .and_then(|v| v.as_str())
-                            .unwrap_or("");
-                        let generated = rew
-                            .get("generatedPath")
-                            .and_then(|v| v.as_str())
-                            .unwrap_or("");
-                        if !workspace.is_empty() && !generated.is_empty() {
-                            rewrite_external_imports = Some(RewriteExternalImportsOptions {
-                                source_path: filename.clone(),
-                                generated_path: generated.to_string(),
-                                workspace_path: workspace.to_string(),
-                            });
-                        }
-                    }
-                }
-            }
-        }
-
-        Svelte2TsxOptions {
-            filename,
-            is_ts_file,
-            mode,
-            accessors,
-            namespace,
-            version,
-            runes: None,
-            emit_jsdoc,
-            rewrite_external_imports,
-        }
-    }
-
-    /// Relaxed comparison for when no `expected-svelte5.ts` exists.
-    ///
-    /// The `expectedv2.ts` file ends with a V4-style class export:
-    ///   `\n\nexport default class Foo extends ...`
-    /// while the Svelte 5 output ends with a V5-style const:
-    ///   `\nconst Foo = __sveltets_2_isomorphic_component(...)`
-    ///
-    /// This function strips both tails and compares just the render body.
-    /// It also removes V5-specific additions not present in V4 expected output,
-    /// and strips V4-specific type assertions (`as {...}`) from the expected output.
-    fn relaxed_compare(actual: &str, expected: &str) -> bool {
-        // Strip component export tail from expected.
-        // V4: `\n\nexport default class ...`
-        // V4 (JSDoc/$$IsomorphicComponent): `*/ export const ...` (the
-        //   `/** @type {$$IsomorphicComponent} */ export const Input__SvelteComponent_`
-        //   form has no preceding newline because the JSDoc and `export` sit
-        //   on the same line — fall back to the `\n/** @template ` marker
-        //   that precedes the `__sveltets_Render` class).
-        // V5: `\nconst ...` or `\nexport const ...`
-        let expect_cut = expected
-            .rfind("\n\nexport default class")
-            .or_else(|| expected.rfind("\nexport const "))
-            .or_else(|| expected.rfind("\n/** @template "))
-            .or_else(|| expected.rfind("\nclass __sveltets_Render"))
-            .or_else(|| expected.rfind("\nconst "));
-        let expect_cut = match expect_cut {
-            Some(pos) => pos,
-            None => return false,
-        };
-        let expected_body = expected[..expect_cut].trim_end();
-
-        // Strip V5-style const export from actual
-        let actual_cut = actual
-            .rfind("\nexport const ")
-            .or_else(|| actual.rfind("\nconst "));
-        let actual_cut = match actual_cut {
-            Some(pos) => pos,
-            None => return false,
-        };
-        let actual_body = actual[..actual_cut].trim_end();
-
-        // Remove V5-specific additions that V4 doesn't have
-        let actual_cleaned = strip_v5_additions(actual_body);
-        // Also strip V5 additions from expected (for V5 expected files)
-        let expected_stripped = strip_v5_additions(expected_body);
-
-        // Direct comparison first
-        if actual_cleaned == expected_body || actual_cleaned == expected_stripped {
-            return true;
-        }
-
-        // Quick whitespace-normalized comparison
-        let actual_ws = normalize_all_whitespace(&actual_cleaned);
-        let expected_ws = normalize_all_whitespace(&expected_stripped);
-        if actual_ws == expected_ws {
-            return true;
-        }
-
-        // Try stripping V4-specific `as {...}` type assertions from expected.
-        // V4 props use `{a: a} as {a?: typeof a}` but V5 just uses `{a: a}`.
-        let expected_cleaned = strip_as_type_assertion(&expected_stripped);
-
-        if actual_cleaned == expected_cleaned {
-            return true;
-        }
-
-        // Try normalizing whitespace in createElement attribute objects.
-        // Template generates `{ "attr"` while expected has `{  "attr"`.
-        let actual_ws_normalized = normalize_attr_whitespace(&actual_cleaned);
-        let expected_ws_normalized = normalize_attr_whitespace(&expected_cleaned);
-
-        if actual_ws_normalized == expected_ws_normalized {
-            return true;
-        }
-
-        // Try collapsing all runs of multiple spaces to single space
-        let actual_collapsed = collapse_spaces(&actual_ws_normalized);
-        let expected_collapsed = collapse_spaces(&expected_ws_normalized);
-
-        if actual_collapsed == expected_collapsed {
-            return true;
-        }
-
-        // Try normalizing props type format (JSDoc vs TS assertion)
-        let actual_props = normalize_props_type(&actual_collapsed);
-        let expected_props = normalize_props_type(&expected_collapsed);
-
-        if actual_props == expected_props {
-            return true;
-        }
-
-        // Normalize svelte:options calls
-        let actual_opts = normalize_svelte_options(&actual_props);
-        let expected_opts = normalize_svelte_options(&expected_props);
-
-        if actual_opts == expected_opts {
-            return true;
-        }
-
-        // Final attempt: normalize return statement
-        // Strip the entire return statement from both and compare just the render body
-        let actual_body = strip_return_statement(&actual_opts);
-        let expected_body = strip_return_statement(&expected_opts);
-
-        if actual_body == expected_body {
-            return true;
-        }
-
-        // Normalize counters, blank lines, and component closing spaces
-        let actual_final =
-            normalize_component_close(&collapse_blank_lines(&normalize_counters(&actual_body)));
-        let expected_final =
-            normalize_component_close(&collapse_blank_lines(&normalize_counters(&expected_body)));
-
-        if actual_final == expected_final {
-            return true;
-        }
-
-        // Normalize template literal strings to regular strings
-        // `text` → "text", `` → ""
-        let actual_tl = normalize_template_literals(&actual_final);
-        let expected_tl = normalize_template_literals(&expected_final);
-
-        if actual_tl == expected_tl {
-            return true;
-        }
-
-        // Final aggressive normalization: normalize all whitespace
-        // (leading indentation, multiple spaces, tabs) to single spaces
-        let actual_normalized = normalize_all_whitespace(&actual_tl);
-        let expected_normalized = normalize_all_whitespace(&expected_tl);
-
-        if actual_normalized == expected_normalized {
-            return true;
-        }
-
-        // Normalize prop shorthand: expand `name,` to `"name":name,` for comparison
-        let actual_props_expanded = expand_prop_shorthand(&actual_normalized);
-        let expected_props_expanded = expand_prop_shorthand(&expected_normalized);
-
-        if actual_props_expanded == expected_props_expanded {
-            return true;
-        }
-
-        // Normalize semicolons
-        let actual_semi = normalize_semicolons(&actual_props_expanded);
-        let expected_semi = normalize_semicolons(&expected_props_expanded);
-
-        if actual_semi == expected_semi {
-            return true;
-        }
-
-        // Strip generic type parameters from function calls for comparison:
-        // `createEventDispatcher<...>()` → `createEventDispatcher()`
-        let actual_generics = strip_call_generics(&actual_semi);
-        let expected_generics = strip_call_generics(&expected_semi);
-
-        if actual_generics == expected_generics {
-            return true;
-        }
-
-        // Strip ignore markers: `/*Ωignore_startΩ*/.../*Ωignore_endΩ*/`
-        let actual_no_ignore = strip_ignore_markers(&actual_generics);
-        let expected_no_ignore = strip_ignore_markers(&expected_generics);
-
-        if actual_no_ignore == expected_no_ignore {
-            return true;
-        }
-
-        // Strip CSS prop wrappers: `...__sveltets_2_cssProp({"key":val})` → `"key":val`
-        // Then re-normalize whitespace
-        let actual_css = normalize_all_whitespace(&strip_css_prop_wrappers(&actual_no_ignore));
-        let expected_css = normalize_all_whitespace(&strip_css_prop_wrappers(&expected_no_ignore));
-
-        if actual_css == expected_css {
-            return true;
-        }
-
-        // Final aggressive fallback: strip *all* whitespace differences
-        // (including spaces inside `{ … }`, around `;`, and across line
-        // boundaries). Used only when every other normaliser still
-        // disagrees on whitespace — typically when the JS reference's
-        // MagicString-position-preserving output differs purely in
-        // padding from our concatenated output.
-        let actual_no_ws = strip_all_whitespace(&actual_css);
-        let expected_no_ws = strip_all_whitespace(&expected_css);
-        if actual_no_ws == expected_no_ws {
-            return true;
-        }
-
-        // Even more permissive: strip whitespace and comments from the
-        // raw cleaned bodies (before all the targeted normalisers above).
-        // Some normalisers leave content behind that can't be recovered,
-        // so this last resort starts from the V5-stripped bodies and
-        // strips comments + whitespace.
-        let actual_raw = strip_all_whitespace(&actual_cleaned);
-        let expected_raw = strip_all_whitespace(&expected_stripped);
-        actual_raw == expected_raw
-    }
-
-    /// Strip *all* whitespace characters AND `//` line + `/* … */` block
-    /// comments. Used as the most permissive fallback in `relaxed_compare`
-    /// so position-preserving padding and comment-preserving differences
-    /// don't fail the comparison when the underlying TSX semantics match.
-    fn strip_all_whitespace(text: &str) -> String {
-        let no_comments = strip_js_comments(text);
-        no_comments.chars().filter(|c| !c.is_whitespace()).collect()
-    }
-
-    /// Remove `//` line comments and `/* … */` block comments. Tries to
-    /// avoid stripping inside string / template literals.
-    fn strip_js_comments(text: &str) -> String {
-        let bytes = text.as_bytes();
-        let len = bytes.len();
-        let mut out = String::with_capacity(len);
-        let mut i = 0;
-        while i < len {
-            let b = bytes[i];
-            if b == b'\'' || b == b'"' || b == b'`' {
-                let q = b;
-                let start = i;
-                i += 1;
-                while i < len && bytes[i] != q {
-                    if bytes[i] == b'\\' && i + 1 < len {
-                        i += 2;
-                        continue;
-                    }
-                    i += 1;
-                }
-                let end = (i + 1).min(len);
-                out.push_str(&text[start..end]);
-                i = end;
-                continue;
-            }
-            if b == b'/' && i + 1 < len {
-                if bytes[i + 1] == b'/' {
-                    while i < len && bytes[i] != b'\n' {
-                        i += 1;
-                    }
-                    continue;
-                }
-                if bytes[i + 1] == b'*' {
-                    i += 2;
-                    while i + 1 < len && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                        i += 1;
-                    }
-                    i = (i + 2).min(len);
-                    continue;
-                }
-            }
-            // For non-ASCII bytes, copy the whole UTF-8 char by slicing
-            // to the next char boundary. (Iterating chars directly
-            // would be cleaner but interferes with the byte-level
-            // string / comment scanner above.)
-            let mut next = i + 1;
-            while next < len && !text.is_char_boundary(next) {
-                next += 1;
-            }
-            out.push_str(&text[i..next]);
-            i = next;
-        }
-        out
-    }
-
-    /// Normalize template literal strings to double-quoted strings.
-    /// Converts `` `text` `` to `"text"` for comparison purposes.
-    /// Only converts simple template literals (no `${...}` interpolations).
-    fn normalize_template_literals(text: &str) -> String {
-        use regex::Regex;
-        // Match template literals that contain no ${...} interpolation
-        let re = Regex::new(r"`([^`$]*)`").unwrap();
-        re.replace_all(text, "\"$1\"").to_string()
-    }
-
-    /// Strip CSS prop wrappers for comparison.
-    /// `...__sveltets_2_cssProp({"--key":val})` → `"--key":val`
-    fn strip_css_prop_wrappers(text: &str) -> String {
-        let needle = "...__sveltets_2_cssProp(";
-        let mut result = text.to_string();
-        while let Some(pos) = result.find(needle) {
-            let after = &result[pos + needle.len()..];
-            // Find the matching closing `)` using brace-aware scanning
-            let mut depth_brace = 0i32;
-            let mut depth_paren = 0i32;
-            let mut in_template = false;
-            let mut end_offset = 0;
-            let bytes = after.as_bytes();
-            for (i, &b) in bytes.iter().enumerate() {
-                if in_template {
-                    if b == b'`' {
-                        in_template = false;
-                    }
-                } else {
-                    match b {
-                        b'`' => in_template = true,
-                        b'{' => depth_brace += 1,
-                        b'}' => depth_brace -= 1,
-                        b'(' => depth_paren += 1,
-                        b')' => {
-                            if depth_paren == 0 && depth_brace == 0 {
-                                end_offset = i;
-                                break;
-                            }
-                            depth_paren -= 1;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            if end_offset > 0 {
-                // Extract inner content (skip outermost `{...}`)
-                let inner = &after[1..end_offset - 1]; // skip `{` and `}`
-                let full_end = pos + needle.len() + end_offset + 1; // +1 for `)`
-                result = format!("{}{}{}", &result[..pos], inner, &result[full_end..]);
-            } else {
-                break;
-            }
-        }
-        result
-    }
-
-    /// Strip ignore markers from text for comparison.
-    /// Removes `/*Ωignore_startΩ*/.../*Ωignore_endΩ*/` sections.
-    fn strip_ignore_markers(text: &str) -> String {
-        let start_marker = "/*\u{03A9}ignore_start\u{03A9}*/";
-        let end_marker = "/*\u{03A9}ignore_end\u{03A9}*/";
-        let mut result = text.to_string();
-        while let Some(start_pos) = result.find(start_marker) {
-            if let Some(end_pos) = result[start_pos..].find(end_marker) {
-                let remove_end = start_pos + end_pos + end_marker.len();
-                result = format!("{}{}", &result[..start_pos], &result[remove_end..]);
-            } else {
-                break;
-            }
-        }
-        result
-    }
-
-    /// Strip generic type parameters from function calls.
-    /// `createEventDispatcher<Type>()` → `createEventDispatcher()`
-    fn strip_call_generics(text: &str) -> String {
-        // Match `identifier<...>(` and remove the `<...>` part.
-        // Need to handle nested angle brackets.
-        let mut result = String::with_capacity(text.len());
-        let chars: Vec<char> = text.chars().collect();
-        let mut i = 0;
-        while i < chars.len() {
-            if chars[i] == '<' && i > 0 && (chars[i - 1].is_alphanumeric() || chars[i - 1] == '_') {
-                // Might be a generic: find the matching `>`
-                let mut depth = 1;
-                let mut j = i + 1;
-                while j < chars.len() && depth > 0 {
-                    if chars[j] == '<' {
-                        depth += 1;
-                    } else if chars[j] == '>' {
-                        depth -= 1;
-                    }
-                    j += 1;
-                }
-                if depth == 0 && j < chars.len() && chars[j] == '(' {
-                    // Valid generic call: skip the `<...>` part
-                    i = j;
-                    continue;
-                }
-            }
-            result.push(chars[i]);
-            i += 1;
-        }
-        result
-    }
-
-    /// Normalize trailing semicolons: strip all trailing semicolons from non-empty lines.
-    /// This handles differences like `import A` vs `import A;`.
-    fn normalize_semicolons(text: &str) -> String {
-        text.lines()
-            .map(|line| {
-                let trimmed = line.trim_end();
-                if trimmed.ends_with(';') {
-                    trimmed.trim_end_matches(';').to_string()
-                } else {
-                    trimmed.to_string()
-                }
-            })
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    /// Expand prop shorthand to full key-value format for comparison.
-    /// `foo,` → `"foo":foo,` (when preceded by `{` or `,`)
-    /// This handles the difference between shorthand and full format in props/slots.
-    fn expand_prop_shorthand(text: &str) -> String {
-        use regex::Regex;
-        // Match shorthand props: identifier followed by comma, preceded by `{` or `,`
-        // `{foo,bar,}` → `{"foo":foo,"bar":bar,}`
-        let re = Regex::new(r"([{,])\s*([a-zA-Z_$][a-zA-Z0-9_$]*)\s*,").unwrap();
-        re.replace_all(text, |caps: &regex::Captures| {
-            let prefix = &caps[1];
-            let name = &caps[2];
-            format!("{}\"{}\":{},", prefix, name, name)
-        })
-        .to_string()
-    }
-
-    /// Aggressively normalize all whitespace for comparison.
-    /// Collapses all runs of whitespace (spaces, tabs) into single spaces,
-    /// and normalizes leading indentation on each line.
-    fn normalize_all_whitespace(text: &str) -> String {
-        text.lines()
-            .map(|line| {
-                let trimmed = line.trim();
-                if trimmed.is_empty() {
-                    String::new()
-                } else {
-                    // Collapse all internal whitespace runs to single space
-                    let mut result = String::new();
-                    let mut last_was_space = false;
-                    for ch in trimmed.chars() {
-                        if ch == ' ' || ch == '\t' {
-                            if !last_was_space {
-                                result.push(' ');
-                                last_was_space = true;
-                            }
-                        } else {
-                            result.push(ch);
-                            last_was_space = false;
-                        }
-                    }
-                    result
-                }
-            })
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    /// Collapse multiple consecutive newlines into a single newline.
-    fn collapse_blank_lines(text: &str) -> String {
-        use regex::Regex;
-        let re = Regex::new(r"\n{2,}").unwrap();
-        re.replace_all(text, "\n").to_string()
-    }
-
-    /// Normalize closing brace + component name patterns.
-    /// `} Component}` → `}Component}`, `} Test}` → `}Test}`
-    fn normalize_component_close(text: &str) -> String {
-        use regex::Regex;
-        // Match `} Identifier}` where Identifier starts with uppercase
-        let re = Regex::new(r"\} ([A-Z][a-zA-Z0-9_]*)\}").unwrap();
-        re.replace_all(text, "}$1}").to_string()
-    }
-
-    /// Normalize component counter numbers in variable names.
-    /// `$$_tseT1C` → `$$_tseT0C`, `$$_tseT1` → `$$_tseT0`, etc.
-    /// This handles cases where the counter starts from different numbers.
-    fn normalize_counters(text: &str) -> String {
-        use regex::Regex;
-        let re = Regex::new(r"\$\$_([a-zA-Z]+)(\d+)").unwrap();
-        re.replace_all(text, "$$_${1}0").to_string()
-    }
-
-    /// Normalize svelte:options createElement calls by removing whitespace differences
-    /// and the entire line containing the call, since the formatting varies between
-    /// V4 and V5 and between bare/expression attributes.
-    fn normalize_svelte_options(text: &str) -> String {
-        use regex::Regex;
-        // Remove the entire svelte:options createElement line
-        let re =
-            Regex::new(r#"[^\n]*svelteHTML\.createElement\("svelte:options"[^\n]*\n?"#).unwrap();
-        re.replace_all(text, "").to_string()
-    }
-
-    /// Strip the return statement and everything after it from the text.
-    /// This allows comparing just the render function body.
-    fn strip_return_statement(text: &str) -> String {
-        if let Some(pos) = text.rfind("\nreturn {") {
-            text[..pos].trim_end().to_string()
-        } else if let Some(pos) = text.rfind("return {") {
-            text[..pos].trim_end().to_string()
-        } else {
-            text.to_string()
-        }
-    }
-
-    /// Normalize props type format: strip both JSDoc and TS type assertions
-    /// so `/** @type {Record<string, never>} */ ({})` and `{} as Record<string, never>`
-    /// both become `{}`.
-    fn normalize_props_type(text: &str) -> String {
-        use regex::Regex;
-        let mut result = text.to_string();
-
-        // Strip JSDoc type assertions: `/** @type {Type} */ (value)` → `value`
-        let jsdoc_re = Regex::new(r"/\*\*\s*@type\s*\{[^}]*\}\s*\*/\s*\(([^)]*)\)").unwrap();
-        result = jsdoc_re.replace_all(&result, "$1").to_string();
-
-        // Strip TS type assertions: `value as Type` → `value`
-        // Be careful to only strip simple `as` assertions, not nested ones
-        let as_re = Regex::new(r"\{\} as Record<string, never>").unwrap();
-        result = as_re.replace_all(&result, "{}").to_string();
-
-        result
-    }
-
-    /// Collapse all runs of multiple spaces to a single space,
-    /// and normalize brace spacing and trailing whitespace for comparison.
-    fn collapse_spaces(text: &str) -> String {
-        use regex::Regex;
-        let re = Regex::new(r" {2,}").unwrap();
-        let mut result = re.replace_all(text, " ").to_string();
-        // Normalize brace spacing
-        result = result.replace("{ }", "{}");
-        // Normalize `{ "` to `{"` (opening brace followed by space and quote)
-        result = result.replace("{ \"", "{\"");
-        // Normalize `) ;` to `);` (space before semicolon after close paren)
-        result = result.replace(") ;", ");");
-        result
-    }
-
-    /// Strip `as {... }` type assertions from the return statement props.
-    ///
-    /// V4 expected output uses patterns like:
-    ///   `props: {a: a} as {a?: typeof a}`
-    /// while V5 just uses:
-    ///   `props: {a: a}`
-    fn strip_as_type_assertion(text: &str) -> String {
-        let mut result = text.to_string();
-
-        // Find patterns like `} as {` and remove ` as {...}`
-        while let Some(pos) = result.find("} as {") {
-            let keep_end = pos + 1; // keep the `}` at pos, remove from ` as {`
-            let brace_start = pos + 5; // position of `{` in ` as {`
-
-            // Find the matching closing brace
-            let mut depth = 0;
-            let mut end_pos = brace_start;
-            for (i, ch) in result[brace_start..].char_indices() {
-                match ch {
-                    '{' => depth += 1,
-                    '}' => {
-                        depth -= 1;
-                        if depth == 0 {
-                            end_pos = brace_start + i + 1;
-                            break;
-                        }
-                    }
-                    _ => {}
-                }
-            }
-
-            if depth == 0 && end_pos > brace_start {
-                // Remove ` as {...}` but keep the original `}`
-                result = format!("{}{}", &result[..keep_end], &result[end_pos..]);
-            } else {
-                break;
-            }
-        }
-
-        // Also handle `as unknown as $$Events` pattern used in event interfaces
-        while let Some(pos) = result.find(" as unknown as ") {
-            // Find what follows (identifier or type)
-            let after = &result[pos + 15..]; // skip " as unknown as "
-            let end = after
-                .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '$')
-                .unwrap_or(after.len());
-            let type_end = pos + 15 + end;
-            result = format!("{}{}", &result[..pos], &result[type_end..]);
-        }
-
-        result
-    }
-
-    /// Strip V5-specific additions from the actual output for relaxed comparison.
-    ///
-    /// Removes patterns like:
-    /// - `, exports: {}`
-    /// - `, exports: /** @type {{...}} */ ({})`
-    /// - `, bindings: ""`
-    /// - `, bindings: __sveltets_$$bindings('')`
-    fn strip_v5_additions(text: &str) -> String {
-        use regex::Regex;
-
-        let mut result = text.to_string();
-
-        // Remove `, exports: ...` using brace-aware matching
-        while let Some(pos) = result.find(", exports: ") {
-            let after = &result[pos + 11..]; // skip ", exports: "
-            // Find the end of the exports value by matching braces
-            if let Some(end_offset) = find_balanced_end(after) {
-                let remove_end = pos + 11 + end_offset;
-                result = format!("{}{}", &result[..pos], &result[remove_end..]);
-            } else {
-                break;
-            }
-        }
-
-        // Remove `, bindings: ""`
-        result = result.replace(", bindings: \"\"", "");
-
-        // Remove `, bindings: __sveltets_$$bindings('...')`
-        let bindings_re = Regex::new(r", bindings: __sveltets_\$\$bindings\('[^']*'\)").unwrap();
-        result = bindings_re.replace_all(&result, "").to_string();
-
-        // Remove `, bindings: __sveltets_$$bindings('...', '...')`
-        let bindings_re2 = Regex::new(r", bindings: __sveltets_\$\$bindings\([^)]*\)").unwrap();
-        result = bindings_re2.replace_all(&result, "").to_string();
-
-        // Remove `children:() => { return __sveltets_2_any(0); },` from component props
-        result = result.replace("children:() => { return __sveltets_2_any(0); },", "");
-        result = result.replace(" children:() => { return __sveltets_2_any(0); },", "");
-
-        // Remove slot creation declaration (V5-specific addition)
-        // `/*Ωignore_startΩ*/;const __sveltets_createSlot = __sveltets_2_createCreateSlot();/*Ωignore_endΩ*/`
-        let slot_decl_re = regex::Regex::new(
-            r"/\*Ωignore_startΩ\*/;const __sveltets_createSlot = __sveltets_2_createCreateSlot[^;]*;\s*/\*Ωignore_endΩ\*/;?"
-        ).unwrap();
-        result = slot_decl_re.replace_all(&result, "").to_string();
-
-        result
-    }
-
-    /// Find the end of a balanced expression starting from the current position.
-    /// Handles nested braces `{}` and parentheses `()`.
-    /// Returns the offset past the last closing delimiter.
-    fn find_balanced_end(text: &str) -> Option<usize> {
-        let mut depth_brace = 0i32;
-        let mut depth_paren = 0i32;
-        let mut in_string = false;
-        let mut string_char = '"';
-        let mut started = false;
-        let mut i = 0;
-        let bytes = text.as_bytes();
-
-        while i < bytes.len() {
-            let ch = bytes[i] as char;
-            if in_string {
-                if ch == string_char && (i == 0 || bytes[i - 1] != b'\\') {
-                    in_string = false;
-                }
-            } else {
-                match ch {
-                    '"' | '\'' | '`' => {
-                        in_string = true;
-                        string_char = ch;
-                    }
-                    '{' => {
-                        depth_brace += 1;
-                        started = true;
-                    }
-                    '}' => {
-                        depth_brace -= 1;
-                        if started && depth_brace == 0 && depth_paren == 0 {
-                            return Some(i + 1);
-                        }
-                    }
-                    '(' => {
-                        depth_paren += 1;
-                        started = true;
-                    }
-                    ')' => {
-                        depth_paren -= 1;
-                        if started && depth_brace == 0 && depth_paren == 0 {
-                            return Some(i + 1);
-                        }
-                    }
-                    ',' if !started && depth_brace == 0 && depth_paren == 0 => {
-                        // End at comma if we haven't started a delimited group
-                        return Some(i);
-                    }
-                    _ => {}
-                }
-            }
-            i += 1;
-        }
-        // If we never found a balanced end, consume everything
-        if started { None } else { Some(text.len()) }
-    }
-
-    /// Normalize whitespace differences in createElement attribute objects.
-    ///
-    /// The template renderer may produce `{ "attr"` or `{  "attr"` (different
-    /// numbers of spaces after the opening brace). This normalizes them to a
-    /// single space so relaxed comparison can succeed.
-    ///
-    /// Also normalizes leading whitespace before `{` and `for(` in template
-    /// contexts, and collapses multiple consecutive spaces into single spaces
-    /// in createElement/component contexts.
-    fn normalize_attr_whitespace(text: &str) -> String {
-        use regex::Regex;
-        // Normalize multiple spaces after `{` in createElement contexts
-        let re = Regex::new(r"\{\s{2,}").unwrap();
-        let result = re.replace_all(text, "{ ").to_string();
-
-        // Normalize multiple spaces to single space (preserving indentation)
-        // This handles whitespace differences in template attribute output.
-        let re2 = Regex::new(r"([^ \t\n])  +").unwrap();
-        let result = re2.replace_all(&result, "$1 ").to_string();
-
-        result
-    }
-
-    /// Build a compact diff snippet showing the first N lines that differ.
-    fn first_diff_snippet(actual: &str, expected: &str, context_lines: usize) -> String {
-        let actual_lines: Vec<&str> = actual.lines().collect();
-        let expected_lines: Vec<&str> = expected.lines().collect();
-        let max_len = actual_lines.len().max(expected_lines.len());
-
-        let diff_line = (0..max_len).find(|&i| {
-            actual_lines.get(i).copied().unwrap_or("")
-                != expected_lines.get(i).copied().unwrap_or("")
-        });
-
-        match diff_line {
-            Some(line_idx) => {
-                let mut out = String::new();
-                out.push_str(&format!("  First difference at line {}:\n", line_idx + 1));
-                let start = line_idx.saturating_sub(1);
-                let end = (line_idx + context_lines).min(max_len);
-                for i in start..end {
-                    let a = actual_lines.get(i).copied().unwrap_or("<missing>");
-                    let e = expected_lines.get(i).copied().unwrap_or("<missing>");
-                    if a == e {
-                        out.push_str(&format!("    {}: {}\n", i + 1, a));
-                    } else {
-                        out.push_str(&format!("  - {}: {}\n", i + 1, e));
-                        out.push_str(&format!("  + {}: {}\n", i + 1, a));
-                    }
-                }
-                out
-            }
-            None => "  (outputs have different trailing content)\n".to_string(),
-        }
-    }
-
-    // =========================================================================
-    // Main test
-    // =========================================================================
-
-    #[test]
-    fn test_svelte2tsx_fixtures() {
-        let samples_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("submodules/language-tools/packages/svelte2tsx/test/svelte2tsx/samples");
-
-        if !samples_dir.exists() {
-            eprintln!(
-                "Skipping: language-tools submodule not available at {:?}",
-                samples_dir
-            );
+//! Prints per-sample status and a final pass-rate summary.
+
+mod common;
+
+use common::TestStatus;
+use common::svelte2tsx::iter_svelte2tsx_outcomes;
+
+#[test]
+fn test_svelte2tsx_fixtures() {
+    let outcomes = match iter_svelte2tsx_outcomes() {
+        Some(o) => o,
+        None => {
+            eprintln!("Skipping: language-tools submodule not available");
             return;
         }
+    };
 
-        let mut passed = 0u32;
-        let mut failed = 0u32;
-        let mut skipped = 0u32;
-        let mut panic_count = 0u32;
-        let mut error_count = 0u32;
-        let mut failures: Vec<String> = Vec::new();
+    let mut passed = 0u32;
+    let mut failed = 0u32;
+    let mut skipped = 0u32;
+    let mut error_count = 0u32;
+    let mut panic_count = 0u32;
+    let mut failures: Vec<String> = Vec::new();
 
-        let mut entries: Vec<_> = fs::read_dir(&samples_dir)
-            .expect("failed to read samples directory")
-            .filter_map(|e| e.ok())
-            .collect();
-        entries.sort_by_key(|e| e.file_name());
-
-        for entry in &entries {
-            let sample_name = entry.file_name().to_string_lossy().to_string();
-            let sample_dir = entry.path();
-
-            // Skip hidden directories
-            if sample_name.starts_with('.') {
-                continue;
+    for outcome in &outcomes {
+        match outcome.status {
+            TestStatus::Passed => {
+                passed += 1;
+                println!("PASS: {}", outcome.name);
             }
-
-            // Skip non-directories
-            if !sample_dir.is_dir() {
-                continue;
+            TestStatus::Failed => {
+                failed += 1;
+                let msg = outcome.message.clone().unwrap_or_default();
+                failures.push(format!("FAIL: {}\n{}", outcome.name, msg));
             }
-
-            // Skip error tests (they expect parse failures)
-            if sample_dir.join("expected.error.json").exists() {
+            TestStatus::Skipped => {
                 skipped += 1;
-                continue;
             }
-
-            // Find the svelte input file
-            let input_path = match find_svelte_file(&sample_dir) {
-                Some(p) => p,
-                None => {
-                    skipped += 1;
-                    continue;
-                }
-            };
-            let svelte_filename = input_path
-                .file_name()
-                .unwrap()
-                .to_string_lossy()
-                .to_string();
-            let input = match fs::read_to_string(&input_path) {
-                Ok(s) => s,
-                Err(_) => {
-                    skipped += 1;
-                    continue;
-                }
-            };
-
-            // Determine expected output file (mirrors JS logic):
-            // - For .v5 samples: always use expectedv2.ts
-            // - For other samples: prefer expected-svelte5.ts, fall back to expectedv2.ts
-            let is_v5_sample = sample_name.ends_with(".v5");
-            let has_svelte5_expected =
-                !is_v5_sample && sample_dir.join("expected-svelte5.ts").exists();
-            let expected_path = if has_svelte5_expected {
-                sample_dir.join("expected-svelte5.ts")
-            } else {
-                sample_dir.join("expectedv2.ts")
-            };
-            if !expected_path.exists() {
-                skipped += 1;
-                continue;
-            }
-            let expected = normalize(&fs::read_to_string(&expected_path).unwrap());
-
-            // Build options from sample name
-            let options = build_options(&sample_name, &sample_dir, &svelte_filename);
-
-            // Run svelte2tsx, catching panics to avoid aborting the whole suite
-            let input_clone = input.clone();
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                svelte2tsx(&input_clone, options)
-            }));
-
-            match result {
-                Ok(Ok(output)) => {
-                    let actual = normalize(&output.code);
-                    if actual == expected {
-                        passed += 1;
-                        println!("PASS (exact): {}", sample_name);
-                    } else if normalize_all_whitespace(&normalize_semicolons(&actual))
-                        == normalize_all_whitespace(&normalize_semicolons(&expected))
-                    {
-                        // Whitespace+semicolons normalized match
-                        passed += 1;
-                        println!("PASS (ws-normalized): {}", sample_name);
-                    } else if has_svelte5_expected && relaxed_compare(&actual, &expected) {
-                        // Relaxed comparison also for svelte5 expected tests
-                        passed += 1;
-                        println!("PASS (relaxed-v5): {}", sample_name);
-                    } else if !has_svelte5_expected && relaxed_compare(&actual, &expected) {
-                        // Relaxed match: render body matches, only component export differs
-                        passed += 1;
-                        println!("PASS (relaxed): {}", sample_name);
-                    } else {
-                        failed += 1;
-                        let diff = first_diff_snippet(&actual, &expected, 5);
-                        failures.push(format!("FAIL: {}\n{}", sample_name, diff));
-                    }
-                }
-                Ok(Err(e)) => {
-                    failed += 1;
-                    error_count += 1;
-                    failures.push(format!("ERROR: {} - {}", sample_name, e));
-                }
-                Err(panic_info) => {
-                    failed += 1;
+            TestStatus::Error => {
+                failed += 1;
+                let msg = outcome.message.clone().unwrap_or_default();
+                if msg.starts_with("PANIC:") {
                     panic_count += 1;
-                    let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
-                        s.to_string()
-                    } else if let Some(s) = panic_info.downcast_ref::<String>() {
-                        s.clone()
-                    } else {
-                        "unknown panic".to_string()
-                    };
-                    failures.push(format!("PANIC: {} - {}", sample_name, msg));
+                    failures.push(format!("PANIC: {} - {}", outcome.name, msg));
+                } else {
+                    error_count += 1;
+                    failures.push(format!("ERROR: {} - {}", outcome.name, msg));
                 }
             }
         }
+    }
 
-        // Print summary
-        println!("\n=== svelte2tsx Fixture Results ===");
-        println!("Passed:  {}", passed);
+    println!("\n=== svelte2tsx Fixture Results ===");
+    println!("Passed:  {}", passed);
+    println!(
+        "Failed:  {} (errors: {}, panics: {})",
+        failed, error_count, panic_count
+    );
+    println!("Skipped: {}", skipped);
+    println!("Total:   {}", passed + failed + skipped);
+
+    if !failures.is_empty() {
+        println!("\nFailure names:");
+        for err in &failures {
+            if let Some(first_line) = err.lines().next() {
+                println!("  {}", first_line);
+            }
+        }
+        println!("\nDetailed failures:");
+        for err in failures.iter().take(50) {
+            println!("  {}", err);
+        }
+        if failures.len() > 50 {
+            println!("  ... and {} more", failures.len() - 50);
+        }
+    }
+
+    let total_tested = passed + failed;
+    if total_tested > 0 {
         println!(
-            "Failed:  {} (errors: {}, panics: {})",
-            failed, error_count, panic_count
+            "\nPass rate: {:.1}% ({}/{})",
+            (passed as f64 / total_tested as f64) * 100.0,
+            passed,
+            total_tested
         );
-        println!("Skipped: {}", skipped);
-        println!("Total:   {}", passed + failed + skipped);
-
-        if !failures.is_empty() {
-            println!("\nFailure names:");
-            for err in &failures {
-                // Just print the first line (name) of each failure
-                if let Some(first_line) = err.lines().next() {
-                    println!("  {}", first_line);
-                }
-            }
-            println!("\nDetailed failures:");
-            for err in failures.iter() {
-                println!("  {}", err);
-            }
-            if failures.len() > 50 {
-                println!("  ... and {} more", failures.len() - 50);
-            }
-        }
-
-        let total_tested = passed + failed;
-        if total_tested > 0 {
-            println!(
-                "\nPass rate: {:.1}% ({}/{})",
-                (passed as f64 / total_tested as f64) * 100.0,
-                passed,
-                total_tested
-            );
-        }
     }
 }
